+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 * Add scale clamping as a percentage of icon/text size for improved UI/UX, with defaults to [0.2, 1.0]
 * Added `hideBackspaceOnEmpty` to automatically hide backspace button when input is empty
 * Added `onTextChanged` to capture input text changed as displayed in the `PhoneTextInput`
+* Add `initialText` property to autofill `PhoneTextInput`
 
 ## [1.0.5] - 2023-02-08